+++ packages/node/src/node/client/ClientNodeInteraction.ts
@@ -14,6 +14,7 @@ import {
     ClientBdxResponse,
     ClientInteraction,
     ClientInvoke,
+    ClientProbeOptions,
     ClientRead,
     ClientSubscribe,
     ClientSubscription,
@@ -156,6 +157,10 @@ export class ClientNodeInteraction implements Interactable<ActionContext> {
      */
     async initBdx(request: ClientBdxRequest = {}, context?: ActionContext): Promise<ClientBdxResponse> {
         return this.#interaction.initBdx(request, context);
+    }
+
+    async probe(options?: ClientProbeOptions): Promise<boolean> {
+        return this.#interaction.probe(options);
     }
 
     get #interaction() {